+++ node/runtime/pangolin/src/pallets/proxy.rs
@@ -4,7 +4,7 @@ use scale_info::TypeInfo;
 // --- paritytech ---
 use frame_support::traits::InstanceFilter;
 use pallet_proxy::Config;
-use sp_runtime::{traits::BlakeTwo256, RuntimeDebug};
+use sp_runtime::RuntimeDebug;
 // --- darwinia-network ---
 use crate::*;
 
@@ -131,7 +131,7 @@ impl Config for Runtime {
 	type ProxyDepositFactor = ProxyDepositFactor;
 	type MaxProxies = MaxProxies;
 	type MaxPending = MaxPending;
-	type CallHasher = BlakeTwo256;
+	type CallHasher = Hashing;
 	type AnnouncementDepositBase = AnnouncementDepositBase;
 	type AnnouncementDepositFactor = AnnouncementDepositFactor;
 	type WeightInfo = ();
